+++ buildSrc/src/main/kotlin/house/greenhouse/greenhouseconfig/gradle/Properties.kt
@@ -6,7 +6,7 @@ object Properties {
     const val MOD_ID = "greenhouseconfig"
     const val MOD_AUTHOR = "Greenhouse Team"
     val MOD_CONTRIBUTORS = listOf("MerchantPug", "Kneelawk")
-    const val DESCRIPTION = "A niche config system used for the Greenhouse Team's mods."
+    const val DESCRIPTION = "A niche config system created for Greenhouse Team mods."
     const val LICENSE = "MPL-2.0"
 
     const val HOMEPAGE = "https://github.com/GreenhouseModding/greenhouse-config"