+++ common/src/test/java/house/greenhouse/greenhouseconfig/test/client/screen/widget/ServerControllable.java
@@ -1,0 +1,6 @@
+package house.greenhouse.greenhouseconfig.test.client.screen.widget;
+
+public interface ServerControllable {
+    boolean isServerControlled();
+    void setServerControlled(boolean value);
+}