+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Initial alpha release.
+- Added `getUnsynced` method, for getting the client config values without any server interference.